migrate user db code into separate module
