+++ forest_lite/server/routers/auth.py
@@ -6,8 +6,8 @@ from fastapi import Depends, HTTPException, status
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from jose import JWTError, jwt
 from pydantic import BaseModel
-from passlib.context import CryptContext
 from typing import Optional
+from forest_lite.server.user_db import verify_password, get_users_db
 
 
 # to get a string like this run:
@@ -19,41 +19,6 @@ ACCESS_TOKEN_EXPIRE_MINUTES = 30
 
 # Security
 oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")
-pwd_context = CryptContext(schemes=["bcrypt"], deprecated=["auto"])
-
-
-def get_password_hash(password):
-    return pwd_context.hash(password)
-
-
-def get_users_db():
-    # Fake users database
-    return {
-        "johndoe": {
-            "username": "johndoe",
-            "full_name": "John Doe",
-            "group": "highway",
-            "email": "johndoe@example.com",
-            "hashed_password": get_password_hash("secret"),
-            "disabled": False,
-        },
-        "alice": {
-            "username": "alice",
-            "full_name": "Alice",
-            "group": "wcssp",
-            "email": "alice@example.com",
-            "hashed_password": get_password_hash("secret2"),
-            "disabled": False,
-        },
-        "bob": {
-            "username": "bob",
-            "full_name": "Bob",
-            "group": "guest",
-            "email": "bob@example.com",
-            "hashed_password": get_password_hash("anonymous"),
-            "disabled": False,
-        }
-    }
 
 
 class Token(BaseModel):
@@ -76,15 +41,8 @@ class User(BaseModel):
     disabled: Optional[bool] = None
 
 
-GUEST_USER = User(username="anonymous")
-
-
 class UserInDB(User):
     hashed_password: str
-
-
-def verify_password(plain_password, hashed_password):
-    return pwd_context.verify(plain_password, hashed_password)
 
 
 def get_user(db, username: str):
